% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gal2lit.R
\name{gal2lit}
\alias{gal2lit}
\title{Gallons to liters}
\usage{
gal2lit(gal)
}
\arguments{
\item{gal}{A single number, in gallons to be converted to liters}
}
\value{
The equivalent, in liters
}
\description{
Gallons to liters
}
\details{
The forula is Liters = 3.785 * Gallons
}
